Generic repository design pattern uygulandı

+++ NorthwindAPI/Entities/Concrete/Product.cs
@@ -10,7 +10,7 @@ namespace Entities.Concrete
         public int ProductId { get; set; }
         public int CategoryId { get; set; }
         public string ProductName { get; set; }
-        public double UnitPrice { get; set; }
-        public int UnitsInStock { get; set; }
+        public decimal UnitPrice { get; set; }
+        public short UnitsInStock { get; set; }
     }
 }
